BTHAB-33: Pass value required to prefill contribution via varaibles
When the url is accessed via a modal popup using the url params from window.location becomes unreliable, since the main page is what is binded to the wondow and not the modal url. to ensure we are always having the right values we get params from the url through PHP and then pass to the JS script.  we also refactor the code for readability